requirements.txt
Flask==3.1.0
Pillow==10.4.0
requests==2.32.3